% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{percent}
\alias{percent}
\title{Transform values to percentage values}
\usage{
percent(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The percentage value of \code{x} with \code{y} digits.
}
\description{
Transform values to percentage values
}
\examples{
percent(0.56733, 2)
percent(0.04757213, 4)
}

